+++ weather-service/src/main/java/com/waither/weatherservice/controller/WeatherController.java
@@ -1,6 +1,6 @@
 package com.waither.weatherservice.controller;
 
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,14 +14,20 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/weather")
+@RequestMapping("/api/v1/weather")
 public class WeatherController {
 
 	private final WeatherService weatherService;
 
-	@PostMapping("/main")
+	@GetMapping("/main")
 	public ApiResponse<MainWeatherResponse> getMainWeather(@RequestBody GetWeatherRequest getWeatherRequest) {
 		return ApiResponse.onSuccess(
 			weatherService.getMainWeather(getWeatherRequest.latitude(), getWeatherRequest.longitude()));
+	}
+
+	@GetMapping("/region")
+	public ApiResponse<String> convertGpsToRegionName(@RequestBody GetWeatherRequest getWeatherRequest) {
+		return ApiResponse.onSuccess(
+			weatherService.convertGpsToRegionName(getWeatherRequest.latitude(), getWeatherRequest.longitude()));
 	}
 }
