+++ weather-service/src/main/java/com/waither/weatherservice/repository/DailyWeatherRepository.java
@@ -1,4 +1,4 @@
-package com.waither.weatherservice.redis;
+package com.waither.weatherservice.repository;
 
 import org.springframework.data.repository.CrudRepository;
