--- weather-service/src/main/java/com/waither/weatherservice/swagger/SwaggerConfig.java
@@ -1,27 +1,0 @@
-package com.waither.weatherservice.swagger;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI waitherAPI() {
-        Info info = new Info()
-                .title("Waither Weather API Docs")
-                .description("Waither Weather Service API 명세서입니다.")
-                .version("1.0.0");
-
-
-        return new OpenAPI()
-                .addServersItem(new Server().url("/"))
-                .info(info);
-    }
-}
